\name{Anno450K}
\alias{Anno450K}
\docType{data}
\title{
Anno450K
}
\description{
HumanMethylation450K Loading Information, generated from HumanMethylation450_15017482_v1-2.csv.
}
\usage{data(Anno450K)}
\format{
  Loading Information, probe and their address on chip.
  \describe{
  	\item{\code{Annotation}}{Mapping between CpG ID and Address on Chip}
	\item{\code{ControlProbe}}{Control Probe information, extracted from very origin Manifest.}
  }
}
\examples{
data(Anno450K)
}
\keyword{datasets}
