\name{CellTypeMeans27K}
\alias{CellTypeMeans27K}
\docType{data}
\title{
Cell-type purified whole blood methylation CpG Sites for use with the RefBaseEWAS cell-type correction function.
}
\description{
This dataset contains 500 CpGs sites from cell type purified whole blood samples.
}
\usage{data(CellTypeMeans450K)}
\format{
  The format is:
  num [1:500, 1:6] 0.901 0.9188 0.8503 0.8913 0.0436 ...
  - attr(*, "dimnames")=List of 2
  ..$ : chr [1:500] "cg00226923" "cg16698623" "cg14102807" "cg14127336" ...
  ..$ : chr [1:6] "CD8T" "CD4T" "NK" "Bcell" ...
}

\examples{
data(CellTypeMeans27K)
}
\keyword{datasets}
