\name{CellTypeMeans450K}
\alias{CellTypeMeans450K}
\docType{data}
\title{
Cell-type purified whole blood methylation CpG Sites for use with the RefBaseEWAS cell-type correction function.
}
\description{
This dataset contains 600 CpGs sites from cell type purified whole blood samples.
}
\usage{data(CellTypeMeans450K)}
\format{
  The format is:
 num [1:600, 1:6] 0.125 0.2 0.178 0.284 0.226 ...
 - attr(*, "dimnames")=List of 2
 ..$ : chr [1:600] "cg25939861" "cg00219921" "cg08777095" "cg04329870" ...
 ..$ : chr [1:6] "CD8T" "CD4T" "NK" "Bcell" ...
}

\examples{
data(CellTypeMeans450K)

}
\keyword{datasets}
