\name{EPIC.manifest.hg19}
\alias{EPIC.manifest.hg19}
\docType{data}
\title{
SNP annotation for EPIC array.
}
\description{
This dataset is added to do filtering on SNP for EPIC array. Data is based on hg19.
}
\usage{data(EPIC.manifest.hg19)}
\format{
'data.frame':   866895 obs. of  41 variables:
$ seqnames                : Factor w/ 26 levels "chr1","chr2",..: 1 1 1 1 1 1 1 1 1 1 ...
$ start                   : int  10525 10848 10850 15865 18827 29407 29425 36603 68849 68889 ...
$ end                     : int  10526 10849 10851 15866 18828 29408 29426 36604 68850 68890 ...
$ width                   : int  2 2 2 2 2 2 2 2 2 2 ...
$ strand                  : Factor w/ 3 levels "+","-","*": 3 3 3 3 3 3 3 3 3 3 ...
$ addressA                : num  21611527 91693541 82663207 2665852 84794291 ...
$ addressB                : num  NA 47784201 3701821 39757192 NA ...
$ channel                 : chr  "Both" "Grn" "Grn" "Red" ...
$ designType              : Factor w/ 2 levels "I","II": 2 1 1 1 2 1 1 2 2 2 ...
$ nextBase                : chr  "G/A" "C" "C" "A" ...
$ nextBaseRef             : chr  "C" "G" "G" "C" ...
$ probeType               : chr  "cg" "cg" "cg" "cg" ...
$ orientation             : Factor w/ 2 levels "down","up": 1 2 2 1 1 1 1 2 1 1 ...
$ probeCpGcnt             : int  3 7 8 2 2 8 7 1 1 0 ...
$ context35               : int  4 12 12 4 3 11 12 1 1 1 ...
$ probeStart              : int  10526 10800 10802 15865 18828 29407 29425 36554 68850 68890 ...
$ probeEnd                : num  10575 10849 10851 15914 18877 ...
$ ProbeSeqA               : chr  "AAACRAAACTACRTTATCCTCTACACAAATTTCRATAATACTCTAAAAAC" "ACACATACTAACACATCAAAATAAAAACATAACACAAACACAAAAAAACA" "ACATACTAACACATCAAAATAAAAACATAACACAAACACAAAAAAACACA" "CCAATAACTAACCACTCTACTAAAATCCATCCACCAAACTAAAAACATCA" ...
$ ProbeSeqB               : chr  "" "ACACATACTAACGCGTCGAAATAAAAACGTAACGCAAACGCAAAAAAACG" "ACATACTAACGCGTCGAAATAAAAACGTAACGCAAACGCAAAAAAACGCG" "CCGATAACTAACCACTCTACTAAAATCCATCCGCCAAACTAAAAACATCG" ...
$ chrmA                   : chr  "chr1" "chr1" "chr1" "chr1" ...
$ begA                    : int  10526 10800 10802 15865 18828 29407 29425 36554 68850 68890 ...
$ flag.A                  : int  16 0 0 16 16 16 16 0 16 16 ...
$ mapQ.A                  : int  0 22 22 17 1 3 3 0 6 6 ...
$ cigarA                  : chr  "50M" "50M" "50M" "50M" ...
$ chrmB                   : chr  "." "chr1" "chr1" "chr1" ...
$ begB                    : int  NA 10800 10802 15865 NA 29407 29425 NA NA NA ...
$ flag.B                  : int  NA 0 0 16 NA 16 16 NA NA NA ...
$ mapQ.B                  : int  NA 22 22 17 NA 3 3 NA NA NA ...
$ cigarB                  : chr  "." "50M" "50M" "50M" ...
$ posMatch                : logi  NA NA NA NA NA NA ...
$ MASK.mapping            : logi  TRUE FALSE FALSE FALSE TRUE TRUE ...
$ MASK.typeINextBaseSwitch: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
$ MASK.rmsk15             : logi  TRUE TRUE TRUE FALSE FALSE FALSE ...
$ MASK.sub35.copy         : logi  TRUE FALSE TRUE TRUE FALSE TRUE ...
$ MASK.sub30.copy         : logi  TRUE TRUE TRUE TRUE FALSE TRUE ...
$ MASK.sub25.copy         : logi  TRUE TRUE TRUE TRUE FALSE TRUE ...
$ MASK.sub40.copy         : logi  TRUE FALSE FALSE TRUE FALSE FALSE ...
$ MASK.snp5.common        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
$ MASK.snp5.GMAF1p        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
$ MASK.extBase            : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
$ MASK.general            : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
}
\examples{
data(EPIC.manifest.hg19)
}
\keyword{datasets}
