\name{EPICSimData}
\alias{EPICSimData}
\docType{data}
\title{
Simulation EPIC beadarray Dataset.
}
\description{
This dataset is available to test ChAMP functions on EPIC array.
}
\usage{data(EPICSimData)}
\format{
List of 3
 $ beta: num [1:780385, 1:16] 0.7382 0.8572 0.8624 0.4492 0.0751 ...
 ..- attr(*, "dimnames")=List of 2
 .. ..$ : chr [1:780385] "cg14817997" "cg16269199" "cg13869341" "cg02404219" ...
 .. ..$ : chr [1:16] "HELA_Sim_1" "HELA_Sim_2" "HELA_Sim_3" "HELA_Sim_4" ...
 $ pd  :'data.frame':    16 obs. of  10 variables:
 ..$ Sample_Name : chr [1:16] "HELA_Sim_1" "HELA_Sim_2" "HELA_Sim_3" "HELA_Sim_4" ...
 ..$ Sample_Plate: chr [1:16] "EPIC2" "EPIC2" "EPIC2" "EPIC2" ...
 ..$ Sample_Group: chr [1:16] "control" "control" "control" "control" ...
 ..$ Pool_ID     : logi [1:16] NA NA NA NA NA NA ...
 ..$ Project     : chr [1:16] "X" "X" "X" "X" ...
 ..$ Sample_Well : chr [1:16] "A01" "B01" "C01" "D01" ...
 ..$ Array       : chr [1:16] "R01C01" "R02C01" "R03C01" "R04C01" ...
 ..$ Slide       : chr [1:16] "200079600019" "200079600019" "200079600019" "200079600019" ...
 ..$ Basename    : chr [1:16] "character(0)" "character(0)" "character(0)" "character(0)" ...
 ..$ filenames   : chr [1:16] "./Demo/200079600019_R01C01" "./Demo/200079600019_R02C01" "./Demo/200079600019_R03C01" "./Demo/200079600019_R04C01" ...
 $ detP: num [1:780385, 1:16] 0 0 0 0 0 0 0 0 0 0 ...
 ..- attr(*, "dimnames")=List of 2
 .. ..$ : chr [1:780385] "cg18478105" "cg14361672" "cg01763666" "cg02115394" ...
 .. ..$ : NULL
}
\examples{
data(EPICSimData)
}
\keyword{datasets}
