\name{EPICv1_Mask}
\alias{EPICv1_Mask}
\docType{data}
\title{
SNP annotation for EPIC v1 array.
}
\description{
This dataset is added to do filtering on SNP for EPIC array. Data is based on hg19.
}
\usage{data(EPICv1_Mask)}
\format{
$ probeID            : ilmn name for each CpG
$ MASK_general_AFR   : General CpG should be masked for AFR population.
...
$ MASK_general       : General CpG should be masked, not for any special population.
}
\examples{
data(EPICv1_Mask)
}
\keyword{datasets}
