\name{EPICv2_Mask}
\alias{EPICv2_Mask}
\docType{data}
\title{
SNP annotation for EPIC v2 array.
}
\description{
This dataset is added to do filtering on SNP for EPIC array. Data is based on hg38.
}
\usage{data(EPICv2_Mask)}
\format{
$ probeID            : ilmn name for each CpG
$ MASK_general_AFR   : General CpG should be masked for AFR population.
...
$ MASK_general       : General CpG should be masked, not for any special population.
}
\examples{
data(EPICv2_Mask)
}
\keyword{datasets}
