\name{hm450.manifest.hg19}
\alias{hm450.manifest.hg19}
\docType{data}
\title{
SNP annotation for 450K array.
}
\description{
This dataset is added to do filtering on SNP for 450K array. Data is based on hg19
}
\usage{data(hm450.manifest.hg19)}
\format{
'data.frame':   485577 obs. of  41 variables:
 $ seqnames                : Factor w/ 26 levels "chr1","chr2",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ start                   : int  15865 18827 29407 29425 29435 68849 69591 91550 135252 370260 ...
 $ end                     : int  15866 18828 29408 29426 29436 68850 69592 91551 135253 370261 ...
 $ width                   : int  2 2 2 2 2 2 2 2 2 2 ...
 $ strand                  : Factor w/ 3 levels "+","-","*": 3 3 3 3 3 3 3 3 3 3 ...
 $ addressA                : num  62703328 27651330 25703424 61731400 26752380 ...
 $ addressB                : num  16661461 NA 34666387 14693326 50693408 ...
 $ channel                 : chr  "Red" "Both" "Red" "Red" ...
 $ designType              : Factor w/ 2 levels "I","II": 1 2 1 1 1 2 2 2 2 2 ...
 $ nextBase                : chr  "A" "G/A" "A" "T" ...
 $ nextBaseRef             : chr  "C" "C" "C" "A" ...
 $ probeType               : chr  "cg" "cg" "cg" "cg" ...
 $ orientation             : Factor w/ 2 levels "down","up": 1 1 1 1 1 1 2 2 2 2 ...
 $ probeCpGcnt             : int  2 2 8 7 7 1 3 1 4 3 ...
 $ context35               : int  4 3 11 12 12 1 2 2 4 5 ...
 $ probeStart              : int  15865 18828 29407 29425 29435 68850 69542 91501 135203 370211 ...
 $ probeEnd                : num  15914 18877 29456 29474 29484 ...
 $ ProbeSeqA               : chr  "CCAATAACTAACCACTCTACTAAAATCCATCCACCAAACTAAAAACATCA" "ACTCRAAATTTACTCAATAAACCRTTCAATATATACAAAAACAATTCCCC" "AAAAAAAACACAATAAAAAACAAACAACAACATTAAAACCCAAAAACACA" "AATCCTAAAACCACACTCAAAAAAAACACAATAAAAAACAAACAACAACA" ...
 $ ProbeSeqB               : chr  "CCGATAACTAACCACTCTACTAAAATCCATCCGCCAAACTAAAAACATCG" "" "GAAAAAAACGCAATAAAAAACGAACGACGACGTTAAAACCCGAAAACGCG" "AATCCTAAAACCGCGCTCGAAAAAAACGCAATAAAAAACGAACGACGACG" ...
 $ chrmA                   : chr  "chr1" "chr1" "chr1" "chr1" ...
 $ begA                    : int  15865 18828 29407 29425 29435 68850 69542 91501 135203 370211 ...
 $ flag.A                  : int  16 16 16 16 16 16 0 0 0 0 ...
 $ mapQ.A                  : int  17 1 3 3 3 6 60 6 0 3 ...
 $ cigarA                  : chr  "50M" "50M" "50M" "50M" ...
 $ chrmB                   : chr  "chr1" "." "chr1" "chr1" ...
 $ begB                    : int  15865 NA 29407 29425 29435 NA NA NA NA NA ...
 $ flag.B                  : int  16 NA 16 16 16 NA NA NA NA NA ...
 $ mapQ.B                  : int  17 NA 3 3 3 NA NA NA NA NA ...
 $ cigarB                  : chr  "50M" "." "50M" "50M" ...
 $ posMatch                : logi  NA NA NA NA NA NA ...
 $ MASK.mapping            : logi  FALSE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.typeINextBaseSwitch: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.rmsk15             : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.sub35.copy         : logi  TRUE FALSE TRUE FALSE FALSE TRUE ...
 $ MASK.sub30.copy         : logi  TRUE FALSE TRUE FALSE FALSE TRUE ...
 $ MASK.sub25.copy         : logi  TRUE FALSE TRUE FALSE TRUE TRUE ...
 $ MASK.sub40.copy         : logi  TRUE FALSE FALSE FALSE FALSE TRUE ...
 $ MASK.snp5.common        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.snp5.GMAF1p        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.extBase            : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general            : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
}
\examples{
data(hm450.manifest.hg19)
}
\keyword{datasets}
