\name{hm450.manifest.pop.hg19}
\alias{hm450.manifest.pop.hg19}
\docType{data}
\title{
SNP annotation for 450K array for different population. Data is based on hg19.
}
\description{
This dataset is added to do filtering on SNP for 450K array. Since different population have different SNP probes, this file is included to provide a comprehensive filtering annotation.
}
\usage{data(hm450.manifest.pop.hg19)}
\format{
'data.frame':   866895 obs. of  67 variables:
 $ seqnames        : Factor w/ 26 levels "chr1","chr2",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ start           : int  10525 10848 10850 15865 18827 29407 29425 68849 68889 69591 ...
 $ end             : int  10526 10849 10851 15866 18828 29408 29426 68850 68890 69592 ...
 $ width           : int  2 2 2 2 2 2 2 2 2 2 ...
 $ strand          : Factor w/ 3 levels "+","-","*": 3 3 3 3 3 3 3 3 3 3 ...
 $ MASK.general.AFR: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.AFR   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.EAS: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.EAS   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.EUR: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.EUR   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.SAS: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.SAS   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.AMR: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.AMR   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.GWD: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.GWD   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.YRI: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.YRI   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.TSI: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.TSI   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.IBS: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.IBS   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.CHS: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.CHS   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.PUR: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.PUR   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.JPT: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.JPT   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.GIH: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.GIH   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.CHB: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.CHB   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.STU: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.STU   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.ITU: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.ITU   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.LWK: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.LWK   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.KHV: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.KHV   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.FIN: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.FIN   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.ESN: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.ESN   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.CEU: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.CEU   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.PJL: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.PJL   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.ACB: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.ACB   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.CLM: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.CLM   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.CDX: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.CDX   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.GBR: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.GBR   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.BEB: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.BEB   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.PEL: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.PEL   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.MSL: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.MSL   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.MXL: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.MXL   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 $ MASK.general.ASW: logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ MASK.snp5.ASW   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
}
\examples{
data(hm450.manifest.pop.hg19)
}
\keyword{datasets}
