\name{Hm450K_Mask}
\alias{Hm450K_Mask}
\docType{data}
\title{
SNP annotation for 450K array.
}
\description{
This dataset is added to do filtering on SNP for 450K array. Data is based on hg19.
}
\usage{data(Hm450K_Mask)}
\format{
$ probeID            : ilmn name for each CpG
$ MASK_general_AFR   : General CpG should be masked for AFR population.
...
$ MASK_general       : General CpG should be masked, not for any special population.
}
\examples{
data(Hm450K_Mask)
}
\keyword{datasets}
