\name{illumina450Gr}
\alias{illumina450Gr}
\docType{data}
\title{
GRange Struct of illumina 450K array.
}
\description{
    This dataset will be used in champ.lasso functions.
}

\usage{data(illumina450Gr)}
\format{
  ..@ seqnames       :Formal class 'Rle' [package "S4Vectors"] with 4 slots
  .. .. ..@ values         : Factor w/ 24 levels "chr16","chr3",..: 1 2 3 4 5 1 4 3 6 7 ...
  .. .. ..@ lengths        : int [1:446018] 1 2 1 1 1 1 1 1 1 1 ...
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
  .. .. ..@ start          : int [1:485512] 53468112 37459206 171916037 91194674 42263294 69341139 28890100 41167802 230560793 23034447 ...
  .. .. ..@ width          : int [1:485512] 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..@ NAMES          : chr [1:485512] "cg00000029" "cg00000108" "cg00000109" "cg00000165" ...
  .. .. ..@ elementType    : chr "integer"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ strand         :Formal class 'Rle' [package "S4Vectors"] with 4 slots
  .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 1 2 1 2 1 2 1 2 1 2 ...
  .. .. ..@ lengths        : int [1:242276] 3 2 2 1 2 1 1 1 3 2 ...
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ elementMetadata:Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
  .. .. ..@ rownames       : NULL
  .. .. ..@ nrows          : int 485512
  .. .. ..@ listData       :List of 27
  .. .. .. ..$ chrom_hg18         : Factor w/ 24 levels "chr1","chr10",..: 8 16 16 1 21 6 8 21 1 7 ...
  .. .. .. ..$ MAPINFO_hg18       : int [1:485512] 52025613 37434210 173398731 90967262 42382451 68410892 28797601 41286959 228627416 20585888 ...
  .. .. .. ..$ chrom_hg38         : Factor w/ 37 levels "chr1","chr1_KI270706v1_random",..: 13 26 26 1 32 8 13 32 1 11 ...
  .. .. .. ..$ MAPINFO_hg38       : int [1:485512] 53434200 37417715 172198247 90729117 42405776 68874422 28878779 41310283 230425047 22838619 ...
  .. .. .. ..$ arm                : Factor w/ 2 levels "p","q": 2 1 2 1 1 2 1 1 2 2 ...
  .. .. .. ..$ probeStrand        : chr [1:485512] "+" "+" "+" "-" ...
  .. .. .. ..$ ensemblID          : chr [1:485512] "ENSG00000103479" "ENSG00000198590" "ENSG00000075420" "ENSG00000143032" ...
  .. .. .. ..$ geneSymbol         : chr [1:485512] "RBL2" "C3orf35" "FNDC3B" "BARHL2" ...
  .. .. .. ..$ geneStrand         : Factor w/ 48 levels "-","-;-","-;-;-",..: 24 24 24 1 24 1 24 1 1 1 ...
  .. .. .. ..$ relationtoGene     : Factor w/ 21 levels "downstream","inside",..: 3 2 2 3 2 2 2 3 2 3 ...
  .. .. .. ..$ distancetoTSS      : chr [1:485512] "-250" "31445" "158618" "-11880" ...
  .. .. .. ..$ nearestGeneBoundary: chr [1:485512] "250" "17782" "158618" "11880" ...
  .. .. .. ..$ feature            : Factor w/ 7 levels "1stExon","3'UTR",..: 6 4 4 5 2 2 1 6 5 7 ...
  .. .. .. ..$ cgi                : Factor w/ 4 levels "island","open sea",..: 4 2 2 4 2 4 4 4 4 1 ...
  .. .. .. ..$ cgiName            : chr [1:485512] NA NA NA NA ...
  .. .. .. ..$ featureCgi         : Factor w/ 28 levels "1stExon - island",..: 24 14 14 20 6 8 4 24 20 25 ...
  .. .. .. ..$ conservedTfsb      : Factor w/ 815 levels "V$AHR_01","V$AHR_01;V$EGR1_01;V$EGR2_01;V$EGR3_01;V$NGFIC_01",..: NA NA NA 578 NA NA NA NA NA NA ...
  .. .. .. ..$ bwaMultiMap        : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ promoterAssociated : Factor w/ 2 levels "0","1": 2 1 1 1 1 1 2 2 1 2 ...
  .. .. .. ..$ asnCN              : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ afrCN              : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 2 1 1 ...
  .. .. .. ..$ amrCN              : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ eurCN              : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 2 1 1 1 ...
  .. .. .. ..$ asnLast4           : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ afrLast4           : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ amrLast4           : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. .. ..$ eurLast4           : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
  .. .. ..@ elementType    : chr "ANY"
  .. .. ..@ elementMetadata: NULL
  .. .. ..@ metadata       : list()
  ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomeInfoDb"] with 4 slots
  .. .. ..@ seqnames   : chr [1:24] "chr16" "chr3" "chr1" "chr8" ...
  .. .. ..@ seqlengths : int [1:24] NA NA NA NA NA NA NA NA NA NA ...
  .. .. ..@ is_circular: logi [1:24] NA NA NA NA NA NA ...
  .. .. ..@ genome     : chr [1:24] NA NA NA NA ...
  ..@ metadata       : list()
}

\examples{
data(illumina450Gr)
}
\keyword{datasets}
