\name{multi.hit}
\alias{multi.hit}
\docType{data}
\title{
multi.hit information.
}
\description{
This dataset contains 9341 multi hit CpGs.
}
\usage{data(multi.hit)}
\format{
  The format is:
  'data.frame': 9341 obs. of  38 variables:
    $ TargetID                   : Factor w/ 485577 levels "cg00000029","cg00000108",..: 30 92 234 246 294 325 358 374 402 418 ...
    $ INFINIUM_DESIGN_TYPE       : Factor w/ 2 levels "I","II": 1 2 1 1 2 1 2 2 2 1 ...
    $ GENOME_BUILD               : int  37 37 37 37 37 37 37 37 37 37 ...
    $ CHR                        : Factor w/ 25 levels "","1","10","11",..: 12 8 9 7 2 21 20 5 20 2 ...
    $ MAPINFO                    : int  54746945 82925333 33318910 81580765 149370974 74508806 58777304 54500832 7339455 149719536 ...
    $ CHROMOSOME_36              : Factor w/ 26 levels "","1","10","11",..: 12 8 9 7 2 21 20 5 20 2 ...
    $ COORDINATE_36              : Factor w/ 484739 levels "","100000057",..: 367785 440537 259149 440391 109936 419166 365930 343920 414240 110329 ...
    $ STRAND                     : Factor w/ 3 levels "","F","R": 3 2 3 2 3 3 3 2 2 3 ...
    $ UCSC_REFGENE_NAME          : Factor w/ 42456 levels "","A1BG","A1CF;A1CF;A1CF",..: 19768 1 1 38813 13130 15618 1 13453 6037 1 ...
    $ UCSC_REFGENE_ACCESSION     : Factor w/ 52670 levels "","NM_000014",..: 35153 1 1 598 51448 2129 1 50622 15082 1 ...
    $ UCSC_REFGENE_GROUP         : Factor w/ 4165 levels "","1stExon","1stExon;1stExon",..: 2970 1 1 2035 2035 943 1 2035 2447 1 ...
    $ UCSC_CPG_ISLANDS_NAME      : Factor w/ 27177 levels "","chr1:10003165-10003585",..: 1 8132 8895 1 1 23921 1 5852 1 400 ...
    $ RELATION_TO_UCSC_CPG_ISLAND: Factor w/ 6 levels "","Island","N_Shelf",..: 1 6 2 1 1 2 1 2 1 2 ...
    $ SNP                        : Factor w/ 139064 levels "","rs10000296",..: 88450 1 93755 1 94630 80342 79504 1 1 89545 ...
    $ SNP_DISTANCE               : Factor w/ 11304 levels "","-1","-1;0",..: 10743 1 2655 1 3352 1938 10243 1 1 5664 ...
    $ snp.hit                    : int  0 0 0 0 1 0 0 0 0 0 ...
    $ bwa.multi.hit              : int  1 1 1 1 1 1 1 1 1 1 ...
    $ analyzed                   : int  0 0 0 0 0 0 0 0 0 0 ...
    $ DNase                      : int  0 0 0 0 0 1 0 0 0 0 ...
    $ h3k4me3                    : int  0 0 0 0 0 1 0 0 0 1 ...
    $ h3k36me3                   : int  0 0 0 0 0 0 0 0 0 0 ...
    $ h3k27ac                    : int  0 0 0 0 0 1 1 0 0 0 ...
    $ h3k4me1                    : int  0 0 0 0 0 0 0 0 0 0 ...
    $ bivalent                   : int  0 0 0 0 0 0 0 0 0 1 ...
    $ h3k27me3                   : int  0 0 0 0 0 0 0 1 0 1 ...
    $ h3k9me3                    : int  0 0 0 0 0 0 0 0 0 0 ...
    $ dmc.constitutive           : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.relapse                : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.T                      : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.MLL                    : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.dic920                 : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.HeH                    : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t119                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t1221                  : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t922                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.iamp                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.undefined              : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.nonrecurrent           : int  NA NA NA NA NA NA NA NA NA NA ...
}

\examples{
data(multi.hit)
}
\keyword{datasets}
