\name{PathwayList}
\alias{PathwayList}
\docType{data}
\title{
PathwayList, used in champ.GSEA()
}
\description{
PathwayList from MsigDB,version 4.28. Contains all pathways would be checked in GSEA process.
}
\usage{data(PathwayList)}
\format{
  A list contains 8567 gene list, each represent on Pathway or Gene Module.
  List of 8567
   $ chr5q23 : chr [1:84] "PTMAP2" "FTMT" "PRR16" "MGC32805" ...
   $ chr16q24 : chr [1:123] "LOC642452" "SNORD68" "FLJ12547" "LOC729942" ...
   $ chr8q24 : chr [1:214] "KIFC2" "LYPD2" "EIF2C2" "ADCY8" ...
   $ chr13q11 : chr [1:22] "LOC645626" "SNX19P2" "SKA3" "ZMYM2" ...
   $ chr7p21 : chr [1:57] "CRS" "RPL36AP26" "RAD17P1" "TAS2R2" ...
   $ chr10q23 : chr [1:134] "SNCG" "LOC389997" "FAM190B" "HPSE2" ...
   ...
}
\examples{
data(PathwayList)
}
\keyword{datasets}
