\name{probe.features.epic}
\alias{probe.features.epic}
\docType{data}
\title{
probe.features.epic
}
\description{
HumanMethylationEPIC probe annotations
}
\usage{data(probe.features.epic)}
\format{
  A data frame with 868565 observations on the following 14 variables.
  \describe{
  	\item{\code{CHR}}{a factor with levels}
	\item{\code{MAPINFO}}{a numeric vector}
  }
}

\examples{
data(probe.features.epic)

}
\keyword{datasets}
