\name{probeInfoALL.epic.lv}
\alias{probeInfoALL.epic.lv}
\docType{data}
\title{
Probe Info Data for use with the BMIQ normalization.
}
\description{
The probe details are formatted here for the BMIQ function.
}
\usage{data(probeInfoALL.epic.lv)}
\format{
  The format is:
List of 3
$ Design : num [1:867531] 2 1 2 2 1 2 2 2 2 1 ...
$ CGI    : num [1:867531] 0 1 0 0 1 1 0 0 0 0 ...
$ probeID: chr [1:867531] "cg07881041" "cg18478105" "cg23229610" "cg03513874" ...
}

\examples{
data(probeInfoALL.epic.lv)

}
\keyword{datasets}
