\name{probeInfoALL.lv}
\alias{probeInfoALL.lv}
\docType{data}
\title{
Probe Info Data for use with the BMIQ normalization.
}
\description{
The probe details are formatted here for the BMIQ function.
}
\usage{data(probeInfoALL.lv)}
\format{
  The format is:
List of 5
 $ typeC    : num [1:485577] 1 1 1 1 1 1 1 1 1 1 ...
 $ Design   : num [1:485577] 2 2 2 2 2 2 2 2 2 1 ...
 $ GeneGroup: int [1:485577] 1 NA 5 NA 6 6 4 1 NA 2 ...
 $ CGI      : num [1:485577] 1 0 0 1 0 1 1 1 1 1 ...
 $ probeID  : chr [1:485577] "cg00000029" "cg00000108" "cg00000109" "cg00000165" ...
}

\examples{
data(probeInfoALL.lv)

}
\keyword{datasets}
