\name{snp.hit}
\alias{snp.hit}
\docType{data}
\title{
CpG site related to SNP sites information.
}
\description{
This dataset contains snp sites located into SNP related sites.
}
\usage{data(snp.hit)}
\format{
  The format is:
  'data.frame':   29481 obs. of  38 variables:
    $ TargetID                   : Factor w/ 485577 levels "cg00000029","cg00000108",..: 16 21 34 65 76 98 119 120 147 160 ...
    $ INFINIUM_DESIGN_TYPE       : Factor w/ 2 levels "I","II": 2 1 2 1 2 2 2 2 2 2 ...
    $ GENOME_BUILD               : int  37 37 37 37 37 37 37 37 37 37 ...
    $ CHR                        : Factor w/ 25 levels "","1","10","11",..: 13 2 2 10 2 13 18 17 3 8 ...
    $ MAPINFO                    : int  23913414 5937253 170490434 57839538 8739981 237027592 178364146 110612894 112068221 90066806 ...
    $ CHROMOSOME_36              : Factor w/ 26 levels "","1","10","11",..: 13 2 2 10 2 13 18 17 3 8 ...
    $ COORDINATE_36              : Factor w/ 484739 levels "","100000057",..: 206129 365300 143074 353467 452543 205188 156567 31100 30980 456465 ...
    $ STRAND                     : Factor w/ 3 levels "","F","R": 2 2 3 3 2 3 2 3 2 2 ...
    $ UCSC_REFGENE_NAME          : Factor w/ 42456 levels "","A1BG","A1CF;A1CF;A1CF",..: 19016 25371 1 37783 30924 812 811 1 1 20318 ...
    $ UCSC_REFGENE_ACCESSION     : Factor w/ 52670 levels "","NM_000014",..: 38959 27879 1 36493 25264 27559 26 1 1 51089 ...
    $ UCSC_REFGENE_GROUP         : Factor w/ 4165 levels "","1stExon","1stExon;1stExon",..: 2035 2035 1 2035 1216 2348 2970 1 1 2496 ...
    $ UCSC_CPG_ISLANDS_NAME      : Factor w/ 27177 levels "","chr1:10003165-10003585",..: 15329 2125 1 1 1 15292 19510 1 2598 1 ...
    $ RELATION_TO_UCSC_CPG_ISLAND: Factor w/ 6 levels "","Island","N_Shelf",..: 2 2 1 1 1 4 6 1 5 1 ...
    $ SNP                        : Factor w/ 139064 levels "","rs10000296",..: 17352 122999 51875 43171 103744 112856 35391 40027 21722 58062 ...
    $ SNP_DISTANCE               : Factor w/ 11304 levels "","-1","-1;0",..: 1351 639 852 852 1410 5828 852 3352 14 508 ...
    $ snp.hit                    : int  1 1 1 1 1 1 1 1 1 1 ...
    $ bwa.multi.hit              : int  0 0 0 0 0 0 0 0 0 0 ...
    $ analyzed                   : int  0 0 0 0 0 0 0 0 0 0 ...
    $ DNase                      : int  0 0 0 0 0 0 1 0 0 0 ...
    $ h3k4me3                    : int  0 0 0 0 0 0 1 0 0 0 ...
    $ h3k36me3                   : int  0 0 0 0 0 0 0 0 0 0 ...
    $ h3k27ac                    : int  0 0 0 0 0 0 1 0 0 0 ...
    $ h3k4me1                    : int  0 0 0 0 0 0 1 0 0 0 ...
    $ bivalent                   : int  0 0 0 0 0 0 0 0 0 0 ...
    $ h3k27me3                   : int  0 0 0 0 0 1 0 0 0 0 ...
    $ h3k9me3                    : int  0 0 0 0 0 0 0 0 0 0 ...
    $ dmc.constitutive           : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.relapse                : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.T                      : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.MLL                    : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.dic920                 : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.HeH                    : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t119                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t1221                  : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.t922                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.iamp                   : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.undefined              : int  NA NA NA NA NA NA NA NA NA NA ...
    $ dmc.nonrecurrent           : int  NA NA NA NA NA NA NA NA NA NA ...
}

\examples{
data(snp.hit)

}
\keyword{datasets}
